Replace size_t (MSVC/Windows specific) with unsigned int.

+++ entsys/entity-manager/entity-attributes-manager/EntityAttributeTypeManager.cpp
@@ -18,7 +18,7 @@ namespace entsys {
     }
 
 
-    size_t EntityAttributeTypeManager::count() const
+    unsigned int EntityAttributeTypeManager::count() const
     {
         return map_of_entity_attribute_types.size();
     }
